% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_celltypes}
\alias{annotate_celltypes}
\title{Annotate cell types based on cluster IDs}
\usage{
annotate_celltypes(
  seurat_obj,
  cluster_ids,
  celltype_labels,
  cluster_column = "seurat_clusters"
)
}
\arguments{
\item{seurat_obj}{Seurat object with cluster assignments}

\item{cluster_ids}{Vector of cluster IDs to annotate}

\item{celltype_labels}{Vector of cell type labels corresponding to cluster_ids}

\item{cluster_column}{Metadata column containing cluster IDs (default: "seurat_clusters")}
}
\value{
Seurat object with added "celltype" metadata
}
\description{
Assigns biological cell type annotations to clusters:
\enumerate{
\item Validates equal length of cluster_ids and celltype_labels
\item Creates cluster-to-celltype mapping
\item Adds "celltype" column to object metadata
\item Handles both numeric and character cluster IDs
}
}
\examples{
\donttest{
obj <- Sphinx:::.sphinx_example_seurat(20)
cls <- levels(obj@meta.data$seurat_clusters)
obj <- annotate_celltypes(obj, cls, paste0("Type", seq_along(cls)))
table(obj@meta.data$celltype)
}
}
