% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.calc_adaptive_window_size}
\alias{.calc_adaptive_window_size}
\title{Adaptive window size by iterative density targeting}
\usage{
.calc_adaptive_window_size(
  x,
  y,
  min_cells = 30L,
  max_cells = 50L,
  range_divisor = 100L,
  max_iter = 100L
)
}
\description{
Adaptive window size by iterative density targeting
}
\keyword{internal}
