% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.calc_stereopy_window_size}
\alias{.calc_stereopy_window_size}
\title{Optimal window size via Stereopy CCD iterative adjustment}
\usage{
.calc_stereopy_window_size(
  x,
  y,
  min_cells = 30L,
  max_cells = 50L,
  range_divisor = 100L,
  max_iter = 100L
)
}
\description{
Optimal window size via Stereopy CCD iterative adjustment
}
\keyword{internal}
