% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.edges_delaunay_deldir_block}
\alias{.edges_delaunay_deldir_block}
\title{Single-block deldir triangulation -> undirected edges}
\usage{
.edges_delaunay_deldir_block(df, max_edge_length = NULL)
}
\description{
Single-block deldir triangulation -> undirected edges
}
\keyword{internal}
