% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.mean_cells_per_window_grid}
\alias{.mean_cells_per_window_grid}
\title{Mean cells per grid window over the full tissue grid}
\usage{
.mean_cells_per_window_grid(x, y, win_size)
}
\description{
Mean cells per grid window over the full tissue grid
}
\keyword{internal}
