% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.sphinx_sci_theme}
\alias{.sphinx_sci_theme}
\title{Publication theme for Sphinx ggplot visuals}
\usage{
.sphinx_sci_theme(base_size = 14, grid = FALSE)
}
\arguments{
\item{base_size}{Base font size (default 14 for SCI figures)}

\item{grid}{Show panel grid (default FALSE)}
}
\description{
Publication theme for Sphinx ggplot visuals
}
\keyword{internal}
