% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.sphinx_soft_diverging}
\alias{.sphinx_soft_diverging}
\title{Candy diverging palette (sky blue - white - candy pink)}
\usage{
.sphinx_soft_diverging(n = 256)
}
\description{
Candy diverging palette (sky blue - white - candy pink)
}
\keyword{internal}
