% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.sphinx_soft_sequential}
\alias{.sphinx_soft_sequential}
\title{Candy sequential palette for continuous heatmaps}
\usage{
.sphinx_soft_sequential(n = 256)
}
\arguments{
\item{n}{Number of colors}
}
\description{
Candy sequential palette for continuous heatmaps
}
\keyword{internal}
