% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\docType{data}
\name{spatial_colors}
\alias{spatial_colors}
\title{Custom color palette for visualizations}
\format{
A character vector of hex color codes.
}
\usage{
spatial_colors
}
\description{
Distinct candy qualitative colors used for cluster UMAP / spatial views in
preprocessing helpers.
}
\keyword{datasets}
