% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{data}
\name{celltype_colors}
\alias{celltype_colors}
\title{Custom color palette for cell type visualizations}
\format{
An object of class \code{character} of length 36.
}
\usage{
celltype_colors
}
\description{
A predefined vector of 36 distinct colors for consistent cell type coloring
}
\keyword{datasets}
