% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.calc_spatial_metrics}
\alias{.calc_spatial_metrics}
\title{Calculate spatial metrics for method selection}
\usage{
.calc_spatial_metrics(df, celltype_col)
}
\arguments{
\item{df}{data.table with spatial data}

\item{celltype_col}{cell type column name}
}
\value{
list of spatial metrics
}
\description{
Calculate spatial metrics for method selection
}
\keyword{internal}
