% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.edges_delaunay}
\alias{.edges_delaunay}
\title{Build edges using Delaunay triangulation}
\usage{
.edges_delaunay(df, max_edge_length = NULL)
}
\arguments{
\item{df}{data.table with spatial data}

\item{max_edge_length}{optional maximum edge length filter}
}
\value{
data.table with edges
}
\description{
Build edges using Delaunay triangulation
}
\keyword{internal}
