% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.edges_knn}
\alias{.edges_knn}
\title{Build edges using kNN method}
\usage{
.edges_knn(df, k)
}
\arguments{
\item{df}{data.table with spatial data}

\item{k}{number of neighbors}
}
\value{
data.table with edges
}
\description{
Build edges using kNN method
}
\keyword{internal}
