% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.edges_radius}
\alias{.edges_radius}
\title{Build edges using radius method}
\usage{
.edges_radius(df, radius, k_cap = 50)
}
\arguments{
\item{df}{data.table with spatial data}

\item{radius}{search radius}

\item{k_cap}{maximum neighbors to consider}
}
\value{
data.table with edges
}
\description{
Build edges using radius method
}
\keyword{internal}
