% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.edges_window}
\alias{.edges_window}
\title{Build edges using window method}
\usage{
.edges_window(df, k, tile)
}
\arguments{
\item{df}{data.table with spatial data}

\item{k}{number of neighbors}

\item{tile}{window size}
}
\value{
data.table with edges
}
\description{
Build edges using window method
}
\keyword{internal}
