% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.round_to_even}
\alias{.round_to_even}
\title{Round to the nearest even integer (minimum 2)}
\usage{
.round_to_even(x)
}
\description{
Round to the nearest even integer (minimum 2)
}
\keyword{internal}
