% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_network.R
\name{.select_method}
\alias{.select_method}
\title{Select optimal method based on spatial metrics}
\usage{
.select_method(m, n)
}
\arguments{
\item{m}{spatial metrics}

\item{n}{number of cells}
}
\value{
selected method name
}
\description{
Select optimal method based on spatial metrics
}
\keyword{internal}
