% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis.R
\name{perform_differential_expression}
\alias{perform_differential_expression}
\title{Perform differential expression analysis}
\usage{
perform_differential_expression(protein_df)
}
\arguments{
\item{protein_df}{Data frame with protein expression and cluster information}
}
\value{
Data frame with differential expression results
}
\description{
Perform differential expression analysis
}
