% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\docType{data}
\name{spatial_colors}
\alias{spatial_colors}
\title{Custom color palette for visualizations}
\format{
An object of class \code{character} of length 36.
}
\usage{
spatial_colors
}
\description{
A predefined vector of 36 distinct colors for consistent plotting
}
\keyword{datasets}
