% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{visualize_results}
\alias{visualize_results}
\title{Visualize clustering results}
\usage{
visualize_results(obj, save_dir = "./")
}
\arguments{
\item{obj}{Processed Seurat object with spatial coordinates}

\item{save_dir}{Output directory (default: "./")}
}
\value{
List containing UMAP and spatial plot objects
}
\description{
Generates and saves two key visualizations:
\enumerate{
\item UMAP projection showing cell clusters
\item Spatial scatter plot showing cluster distribution in tissue context
}
}
\examples{
\donttest{
obj <- Sphinx:::.sphinx_example_seurat(25)
obj <- annotate_celltypes(obj,
  cluster_ids = levels(obj@meta.data$seurat_clusters),
  celltype_labels = paste0("Type", seq_along(levels(obj@meta.data$seurat_clusters)))
)
td <- tempdir()
plots <- visualize_results(obj, save_dir = td)
names(plots)
}
}
