% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{visualize_results}
\alias{visualize_results}
\title{Visualize clustering results}
\usage{
visualize_results(obj, save_dir = "./")
}
\arguments{
\item{obj}{Processed Seurat object with spatial coordinates}

\item{save_dir}{Output directory (default: "./")}
}
\value{
List containing UMAP and spatial plot objects
}
\description{
Generates and saves two key visualizations:
\enumerate{
\item UMAP projection showing cell clusters
\item Spatial scatter plot showing cluster distribution in tissue context
}
}
